+++ laktory/resourcesengines/pulumi/secretscope.py
@@ -1,0 +1,63 @@
+import pulumi
+import pulumi_databricks as databricks
+from laktory.resourcesengines.pulumi.base import PulumiResourcesEngine
+from laktory.models.secrets.secretscope import SecretScope
+
+from laktory._logger import get_logger
+
+logger = get_logger(__name__)
+
+
+class PulumiSecretScope(PulumiResourcesEngine):
+
+    @property
+    def provider(self):
+        return "databricks"
+
+    def __init__(
+            self,
+            name=None,
+            secret_scope: SecretScope = None,
+            opts=None,
+    ):
+        if name is None:
+            name = f"secret-scope-{secret_scope.name}"
+        super().__init__(self.t, name, {}, opts)
+
+        opts = pulumi.ResourceOptions(
+            parent=self,
+            # delete_before_replace=True,
+        )
+
+        keyvault_metadata = None
+        if secret_scope.keyvault_metadata:
+            keyvault_metadata = databricks.SecretScopeKeyvaultMetadataArgs(
+                dns_name=secret_scope.keyvault_metadata.dns_name,
+                resource_id=secret_scope.keyvault_metadata.resource_id,
+            )
+
+        self.secret_scope = databricks.SecretScope(
+                f"secret-scope-{secret_scope.name}",
+                backend_type=secret_scope.backend_type,
+                keyvault_metadata=keyvault_metadata,
+                name=secret_scope.name,
+                opts=opts,
+            )
+
+        for s in secret_scope.secrets:
+            databricks.Secret(
+                f"secret-{secret_scope.name}-{s.key}",
+                key=s.key,
+                string_value=s.value,
+                scope=self.secret_scope.name,
+                opts=opts,
+            )
+
+        for p in secret_scope.permissions:
+            databricks.SecretAcl(
+                f"secret-scope-acl-{secret_scope.name}-{p.principal}",
+                permission=p.permission,
+                principal=p.principal,
+                scope=self.secret_scope.name,
+                opts=opts,
+            )